Add Release info to header
