+++ src/shared/components/layouts/UniProtFooter.tsx
@@ -2,14 +2,8 @@ import { memo, HTMLAttributes } from 'react';
 import { Link } from 'react-router-dom';
 import cn from 'classnames';
 import { CitedIcon, EnvelopeIcon, ExternalLink } from 'franklin-sites';
-import { Method } from 'axios';
-
-import useDataApi from '../../hooks/useDataApi';
-
-import apiUrls from '../../config/apiUrls';
 
 import { Location, LocationToPath } from '../../../app/config/urls';
-import { Namespace } from '../../types/namespaces';
 
 import './styles/footer.scss';
 
@@ -29,6 +23,7 @@ import SERILogo from '../../../images/seri-logo.png';
 
 import ElixirCDRLogo from '../../../images/elixir-cdr.png';
 import CTSLogo from '../../../images/core-trust-seal-logo.png';
+import ReleaseInfo from './ReleaseInfo';
 
 const FooterConsortium = () => (
   <div className="consortium">
@@ -65,30 +60,6 @@ const FooterConsortium = () => (
   </div>
 );
 
-const fetchOptions: { method: Method } = { method: 'HEAD' };
-const Release = () => {
-  // TODO: replace with statistics endpoint
-  const { headers } = useDataApi(
-    `${apiUrls.search(Namespace.uniprotkb)}?query=*&size=0`,
-    fetchOptions
-  );
-  const releaseDate = headers?.['x-release']
-    ? new Date(headers['x-release'])
-    : undefined;
-
-  if (!releaseDate) {
-    return <>Loading release information</>;
-  }
-
-  return (
-    <Link to="/">
-      {/* TODO: don't use release number as date, might be different */}
-      Release {releaseDate.getFullYear()}_
-      {`${releaseDate.getMonth() + 1}`.padStart(2, '0')}
-    </Link>
-  );
-};
-
 const FooterCopyrightAndMisc = () => (
   <div className="copyright-misc">
     <p>
@@ -101,7 +72,7 @@ const FooterCopyrightAndMisc = () => (
       <Link to="/help/privacy">Privacy Notice</Link>
     </p>
     <p>
-      <Release /> | <Link to="/">Statistics</Link>
+      <ReleaseInfo />
     </p>
   </div>
 );
